requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
statsmodels>=0.14.1
matplotlib>=3.8.0
plotly>=5.22.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.2
pydantic>=2.6.0
optuna>=3.6.0

# Temporal Fusion Transformer stack
torch>=2.2.0
pytorch-lightning>=2.2.0
pytorch-forecasting>=1.0.0

# LangChain base
langchain>=0.2.10
langchain-openai>=0.1.17

# IO
openpyxl>=3.1.2
